+++ ci/release_notes.md
@@ -1,1 +1,2 @@
 - Report `crate_version` on stderr in concourse operations
+- Add `ls` command - `cepler ls -e <environment>` lists all files tracked by the current config.